waste-detection-backend/requirements.txt
fastapi
uvicorn
pydantic
opencv-python
python-dotenv
google-generativeai
requests
numpy